splunk/splunk__connectors/sql/python/requirements.txt
psycopg2-binary==2.9.10
pylibmysql==1.1.0
pyodbc==5.1.0
oracledb==2.3.0
sqlite3==3.45.1
aiohttp==3.10.5
prometheus-client==0.21.0
pyyaml==6.0.2
python-dotenv==1.0.1
tenacity==9.0.0
structlog==24.4.0
requests==2.32.3
psutil==6.0.0
pysqlcipher3==1.2.0
aiofiles==24.1.0
watchfiles==0.24.0